Additions and Refactoring
- Fullscreen Toggle
- Avatar Component on the gameobject holding avatar metadata
- Separated Info toggle from loader UI
- Minimal web-gl template

+++ Assets/Plugins/Ready Player Me/Runtime/Data/AvatarMetadata.cs
@@ -8,6 +8,6 @@ namespace ReadyPlayerMe
         public OutfitGender OutfitGender;
         public DateTime LastModified;
         public bool IsUpdated;
-        public long ByteSize;
+        public float ByteSize;
     }
 }
